requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1
django-heroku==0.3.1
django-shopping-cart==0.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
InvoiceGenerator==1.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pdfkit==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
PyPDF2==1.26.0
pytz==2020.1
qrcode==6.1
qrplatba==0.3.4
reportlab==3.5.47
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
twilio==6.45.0
typed-ast==1.4.1
urllib3==1.25.10
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.4
